+++ VCPU/Scripts/VCPU/Opcodes/OpcodeActions/Program/HostCallInstruction.cs
@@ -5,12 +5,12 @@ namespace VirtualCPU.Opcodes
     public class HostCallInstruction : OpcodeInstruction
     {
         public string Name => "HOSTCALL";
-        public bool Accept(byte opcode) => opcode == (byte)OpCodes.HOSTCALL;
+        public bool Accept(int opcode) => opcode == (int)OpCodes.HOSTCALL;
 
-        public void Act(VCPU vCpu, byte opcode, Action<string> crashHandle)
+        public void Act(VCPU vCpu, int opcode, Action<string> crashHandle)
         {
-            byte libraryId = vCpu.Program[vCpu.ProgramCounter + 1];
-            byte callId    = vCpu.Program[vCpu.ProgramCounter + 2];
+            int libraryId = vCpu.Program[vCpu.ProgramCounter + 1];
+            int callId    = vCpu.Program[vCpu.ProgramCounter + 2];
             vCpu.HostCallDispatcher.Dispatch(vCpu, libraryId, callId, crashHandle);
             vCpu.SetProgramCounter(vCpu.ProgramCounter + 3);
         }
